Split AssociatedData file into separate classes
