+++ src/cpptrajfiles
@@ -140,7 +140,7 @@ COMMON_SOURCES= \
         Analysis_Wavelet.cpp \
         ArgList.cpp \
         Array1D.cpp \
-        AssociatedData.cpp \
+        AssociatedData_NOE.cpp \
         Atom.cpp \
         AtomMap.cpp \
         AtomMask.cpp \
